Write initial signature for CopyLoader
The `CopyLoader` class will be the backbone of this package, taking the
user's specifications and performing the desired load. Before code can
be written, the desired behavior of the loader must be well-understood;
to that end, I start by writing a signature with fairly detailed
documentation. More documentation will be needed ahead of release, but
this is a good start and puts me in position to start writing code.